+++ spacepackets/__init__.py
@@ -1,7 +1,6 @@
-__version__ = "0.16.0"
+__version__ = "0.17.0"
 
 import logging
-
 from spacepackets.ccsds import (
     SpacePacketHeader,
     SpacePacket,